Extract local Model scopes to Eloquent Builder traits
